backend/pyproject.toml
[project]
name = "onchain-rca-workbench-backend"
version = "0.1.0"
description = "Post-incident on-chain RCA workbench backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "boto3>=1.34",
  "botocore>=1.34",
  "web3>=6.15",
  "eth-account>=0.11",
  "eth-keys>=0.5",
  "temporalio>=1.5",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "jsonschema>=4.22",
  "pyyaml>=6.0",
  "playwright>=1.45",
  # Runtime dependencies required by vendor/txanalyzer/scripts/pull_artifacts.py.
  "pandas>=1.3.0",
  "requests>=2.25.1",
  "tqdm>=4.60.0",
  "typing-extensions>=4.0.0",
  "pytest>=8.0",
  "pytest-asyncio>=0.23"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
